% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinsExtras-package.R
\docType{package}
\name{pinsExtras-package}
\alias{pinsExtras}
\alias{pinsExtras-package}
\title{pinsExtras: Additional boards for pins}
\description{
Extra board implementations for the \code{pins} package.
}
\author{
\strong{Maintainer}: Kevin Dalton \email{_._@example.com}

Authors:
\itemize{
  \item Kevin Dalton \email{_._@example.com}
}

}
\keyword{internal}
