% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinsExtras-package.R
\docType{package}
\name{pinsExtras-package}
\alias{pinsExtras}
\alias{pinsExtras-package}
\title{pinsExtras: Additional boards for pins}
\description{
Extra board implementations for the \code{pins} package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daltonkw/pinsExtras}
  \item Report bugs at \url{https://github.com/daltonkw/pinsExtras/issues}
}

}
\author{
\strong{Maintainer}: Kevin Dalton \email{daltonkw@users.noreply.github.com}

Authors:
\itemize{
  \item Kevin Dalton \email{daltonkw@users.noreply.github.com}
}

}
\keyword{internal}
