% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinsExtras-package.R
\docType{package}
\name{pinsExtras-package}
\alias{pinsExtras}
\alias{pinsExtras-package}
\title{pinsExtras: Additional boards for pins}
\description{
Extra board implementations for the \code{pins} package.
}
\author{
\strong{Maintainer}: Your Name \email{first.last@example.com}

}
\keyword{internal}
